api/requirements.txt
django==4.1.3
django_extensions==3.2.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
# drf-schema-adapter==3.0.0
git+https://github.com/dameyerdave/drf-schema-adapter@filter_to_filterset#egg=drf-schema-adapter
drf-schema-adapter
markdown==3.4.1
friendlylog==1.0.2
gunicorn==20.1.0
uvicorn==0.20.0
psycopg2==2.9.5
pyyaml==6.0
rdkit==2022.9.2
pandas==1.5.2

